Make truffle-compile better for bundling
- There is no uncaughtException listener in bundled land, it looks
like, so removing the listener that solc adds fails when bundled. To
get around this, add a conditional check to ensure the listener exists.

- F this listener noise. solc doesn’t use the listener anyway (it’s
added by emscripten), so stop trying to manage its existence “just in
case”

(FYI, for background: This listener needs to be removed in a NodeJS
environment because the listener will rethrow all errors making output
really nasty. We don’t need that behavior, hence why the listener is
removed immediately.)