Update on Assert.sol to avoid warnings with "truffle test"
Using "A" as a name for a constant in Assert.sol triggered a warning on each function that was using "A" as parameter during the execution of the "truffle test" command (and there were quite many).
Note : These warnings do not appear on oldest version of node and npm. For example, no warning with node version v6.11.2 and npm v3.10.10 but a warning with npm 5.3.0 and node 8.4.0. The easiest way to remove the warning is to rename this "A" constant.