Add browser compiler supplier (browser: true)
Add browser loading strategy that uses VersionRange to download truffle -- but instead of caching it on disk, it caches it in a script tag. More work can be done here to support all downloaded versions. Currently this supplier will only support the first ersion downloaded, and then no others for the rest of the session.