Normalize the source paths so they match when reading artifact files.
This is a bandaid solution that "works for me".  The problem is that the path written to the build.json files is not internally consistent on Windows (sometimes I get `/C/Path/to/file.sol` and sometimes I get `C:\Path\to\file.sol` and it is also not consistent with what `glob` returns (`c:/Path/to/file.sol`).

This is a terrible bandaid fix that addresses my specific instance of this problem, but I'm sure doesn't work for all scenarios.