Adding a default configuration to the Mainnet
As a solution to the #4482 issue, I have provided this patch.

The recommended configuration for the Mainnet should be provided in the truffle-conf.js. After upgrading to EIP-1559, the Mainnet reports strange errors when deploying and interacting with contracts via Truffle/Infura that do not use type 2 transactions.  For example:
 For example.

- -32000 err: max fee per gas less than block base fee
- Error: cannot estimate gas; transaction may fail or may require manual gas limit

So, I recommend to provide the type 2 transaction parameter in the default configuration.

I also provided a copy of the Rinkeby configuration, since it seems that Ropsten is not popular and Opensea's test network is based on Rinkeby.

In addition, I added the HDWalletProvider's wallet serial number to the configuration. This allows for separate deployments to Mainnet and Rinkeby/Ropsten with different addresses in same mnemonic.

If there is anything wrong with this patch, please let me know and I will improve it. I am very happy to contribute to the Truffle project.