Fix the debugger fails to step over a function
This PR fixes that the debugger fails to stepg over a function when at
the declaration of the function as reported in #4561.

The issue is that the debugger will step into instead of stepping
over the function if at the declarion of an inner function, for
example, function f calls function g, then the debugger will fail to
step over when we are at the declaration of function g.

The root cause is that the stepover() of debugger continues stepping if
1. the function depth is greater or
2. starting lines of the source code are the same.
However, after stepping into the function, the function depth remains
the same as it at the function declaration but the starting line of
source code changes. As a result, stepover() stops stepping.

Also, it is a mutliline statement when we are at the function declaration.

Therefore, the solution to this issue is that
1. the condition of function depth remains unchanged, and
2. the condition of the same starting lines is changed to
   if multiline, continue stepping if starting or ending lines are different
   if not multiline, step to the next logical item

I tested the changes with StactTraceTest.sol in stacktrace-test for the
case that all caller and callee are in the same file and Outer.sol in
inner-outer for the case that the caller and callee are in different files.

Note that there are many multiline source range staements in addition to a
function declaraion, such as for-loop and while-loop etc. This PR only
addresses the case of stepping over a function.