Fix the failure of debugger stepover() at the declaration of a function
This PR fixes the failure of stepover() at the declaration of a function reported in
issue #4561. The issue reports failure with multiline source range. This PR does not
address other multiline cases such as for-loop etc. However, this PR does not change
their behaviors either.

The function depth will increase by 1 when entering a function and decrease by 1
when leaving the function.  Hence, the function depth reveals whether we are inside
or outside of a function.  The logic is the following:
if we are at the declaration of a function, keep stepping if the function depth
is greater than or equal to the starting function depth.
For example, if the function depth is n at the function declaration, then we keep
stepping until the function depth becomes n-1 or the transaction completes.
As a result of this approach, we are past the end of a function.

The unittest includes stacktrace-test, inner-outer, super-call-test, preexisting-test
etc in solidity-test-cases to cover various sol versions.