Printout end of source code if the source range is past the end of source code
This PR is to address the crash reported in issue #4616.

The root cause of the issue is that an instruction's byte-offset to the
start of the range in the source code is past the end of source code.
This is impossible since the byte-offset should always be within the
range of the source code. Then the debugger sets the source range to
null for such an instruction since it does not exist in the source code.
And the null source range causes crash when printing the source code.
However, all instructions along with their attributes including the
byte-offset are given by the source map which is provided by the
compiler. Hence, this issue is caused by the invalid input from the
compiler.

However, the debugger needs to be protected from crash with such an invalid
input or even does something more than just not crash. Two options are
considered when such an instruction is detected. The first option is to
issue a warning message and not print any source code. Hence, there will
be no crash. But this option is simply to protect debugger from crash.
The second option is to print part of the source code, such as the end of
source code, in addition to the warning message. The reason to choose the
end of source code because the byte-offset is past the end of source code.
The sencond option is selected because it protects the debugger from crash,
issues a warning message, and provides some information of the source code.

A note about implementation:
When an invalid byte-offset of an instruction is detected, we only set its
source range to the end of source code for the purpose of displaying source
code. Other attributes of the instruction including the invalid byte-offset
are remained intact. The warning message is issued when printing the source
code.