Keep the input of printState intact
This PR is to keep the input of printState intact.

In order to keep the input intact, I clone the input range and set the lines of the cloned range with the end of source code when the byte-offset is past the end of source code.

Also I use array.length instead of Object.keys() for the end of source code.