Allow the truffle dashboard to disable the process button after click
This PR allows the truffle dashboard to disables the "PROCESS" button and to display
"PROCESSING..." after click the "PROCESS" button as reuqested in issue #4758.

The truffle dashboard displays two buttons: PROCESS and REJECT in the
"INCOMING REQUESTS" page. The user can click one of the buttons to process or reject
the request.  But, after click the "PROCESS" button, it may take some time for the
transaction to complete. In particular, when involving HD wallets, the delay could be
significant.  During the delay period, the truffle dashboard shall disable the buttons
and change the text to "PROCESSING...".