Compiler plugins returns compiler object
This PR is for compiler plugin to return the plugin compiler object.

Currently, the truffle compiler plugin returns the caller with plugin module and
definition and the caller needs to determine the compiler object in order to execute
compilation.  From design view point, it might make more sense for truffle compiler
plugin to return the compiler object instead.  In this way, the caller can invoke
the object to execute compilation without knowing the plugin module and definition.