WIP: interesting!
So it seems config doesn't invalidate the cache because webpack redefines require.
this means that production usage of @truffle/config uses node require, and only loads
truffle-config once per process invocation.

using that, we can modify the config provider thunk resolution to be the result of
calling provider() on the thunk so that provider is only instanciated once.

David mentioned there may be some nonce issue with hdwallet-provider, or its use
in one of its subproviders...