compile-solidity-tests: make solc a devDep to improve test fragility
The *compiles w/ local path* test referenced a file in a node_modules
folder that was already subject to moving due to lerna hoisting
behavior. This uncertainty makes the test unreliable!

This PR adds a pinned version of solc as a devDep, ensuring the file
will be available for testing purposes using `require.resolve`.